more files 8

+++ Source/Life.cs
@@ -1,7 +1,7 @@
 using Godot;
 using System;
 
-public partial class Life(double initialValue = 1.0) : RefCounted
+public partial class Life(int initialValue = 1) : RefCounted
 {
     [Signal]
     public delegate void DeathEventHandler();
@@ -9,19 +9,19 @@ public partial class Life(double initialValue = 1.0) : RefCounted
     [Signal]
     public delegate void ChangeEventHandler();
 
-    public double Value => _value;
-    public double MaxValue => _maxValue;
+    public int Value => _value;
+    public int MaxValue => _maxValue;
 
     public bool IsDead => isDead;
 
-    private double _value = initialValue;
-    private double _maxValue = initialValue;
+    private int _value = initialValue;
+    private int _maxValue = initialValue;
 
     private bool isDead = false;
 
     public void Hit(double damage)
     {
-        _value = Math.Max(0.0, _value - damage);
+        _value = (int)Math.Max(0, _value - damage);
         EmitSignal(SignalName.Change);
 
         if (_value <= 0.0 && !isDead)
